# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: SeDoc API
  version: 1.0.0
  description: Enterprise Document Management System
  contact:
    email: sulthansha@raabyt.com
  license:
    name: Proprietary
servers:
  - url: https://api.vaultdms.io/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Session/user bearer token (JWT) issued by /auth/login. Used by
        interactive clients.
    apiKeyAuth:
      type: http
      scheme: bearer
      description: >
        Service-to-service SeDoc API key. Send as
        `Authorization: Bearer vdms_…` (NOT an X-API-Key header). Keys are
        tenant-scoped and scope-gated — e.g. documents:read, documents:write,
        upload, integrations:read.
...
paths:
...
  /admin/settings:
    get:
      tags: [Admin]
      summary: Read the tenant's feature flags (user-facing wrapper)
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                properties:
                  ai_enabled:        { type: boolean }
                  advanced_workflow: { type: boolean }
                  sso_enabled:       { type: boolean }
                  e_signatures:      { type: boolean }
                  custom_branding:   { type: boolean }
                  api_access:        { type: boolean }
                  data_rooms:        { type: boolean }
    put:
      tags: [Admin]
      summary: Update feature flags (owner only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ai_enabled:        { type: boolean }
                advanced_workflow: { type: boolean }
                sso_enabled:       { type: boolean }
                e_signatures:      { type: boolean }
                custom_branding:   { type: boolean }
                api_access:        { type: boolean }
                data_rooms:        { type: boolean }
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object